WIP - Move WebDriverServer classes into Browser.
The code previously had a slightly odd setup where the Browser class
for WebDriver implementations was mostly using a WebDriverServer
subclass for the actual functionality, but that class and all its
subclasses were defined in a single file rather than in a
browser-specific file. This was used to put the Executor class in
charge of running the driver binary for wdspec tests, whilst the
Browser class was replaced with a dummy implementation.

It makes a lot more sense to have the WebDriver functionality directly
in the Browser class for those cases where it's used. This localises
browser-specific code to a single file rather than spreading it over
several files. It also allows wdspec tests to work like other tests
where the Browser instance owns the driver binary rather than the
executor owning the binary. That also means we don't require
boilerplate executorfoo.py files for each foo that can run wdspec
tests.

Since processes are now always owned by the browser class which lives
on the main thread, output handling is also simplified.