[WebCodecs] Implement keyframe enforcement for flush, configure.
This change checks that the first encoded chunk after a flush() or
configure() is a keyframe. For AV1, VP8, VP9, and H.264 the
bitstream is analyzed to ensure the packet is actually a key frame.

Since Chromium doesn't have a Theora or xHE-AAC parser, no
analysis is done for those codecs beyond checking that the type
field is set to 'key'. On platforms without libvpx or libaom, we
will also skip bitstream analysis.

I originally wrote this change using the media::Vp8Parser and
media::Vp9Parser classes, but since they are not currently used
on Android that brings in 16k+ of binary size. As such I've
switched to using libvpx for bitstream analysis.

This change also adds AV1 and VP8 to our list of tested codecs.

Fixed: 1213680
Change-Id: Ibd8fb9daeb6519c8a72aec66bdc4711261113e28