[LayoutNG] Multicol with OOF contained by inline ancestor
An OOF with an inline container and a multicol containing block
crashed due to two different reasons:

1. A DCHECK that made sure that the block size was complete for a
given element was hit when  NGOutOfFlowLayoutPart::
ComputeInlineContainingBlocks() was called inside a fragmentation
context. To fix this, skip calling ComputeInlineContainingBlocks()
during fragmentation and wait to calculate this once the OOF reaches
the fragmentation context root.

2. The containing block of the inline element was detected to be
the LayoutFlowThread rather than the multicol container because
LayoutObject::FindNonAnonymousContainingBlock() checks
IsAnonymousBlock() when finding the containing block, which can be
a LayoutFlowThread. However, a LayoutFlowThread should not be used
as a containing block, so skip over these when finding the correct
containing block.

This results in different behavior in both legacy and NG than what
legacy produced previously. In the added wpt test, legacy would
break the abspos into two columns. However, with the new changes, the
multicol container is treated as the containing block, and the abspos
does not break. This matches Firefox's behavior, as well.

Bug: 1079031
Change-Id: I90dc78fc541abf2a93a62da808542787a3ad2e6f