Missing unblock of load event in display:none iframe
If we inserted an object into a display:none iframe during parsing after
we had updated the layout tree, we would not unblock the load event
because we did not unblock when the layout tree update was a no-op. This
typically happens because inserting elements into a non-rendered subtree
which is up-to-date does not require a subsequent update.

Makes the html-parser-threaded.html performance test start working
again.

Bug: 1216353
Change-Id: I3e5b0390c45b9e0411fe10fb22a686f39791a66f