Add a WPT to check for standards mode plugin document
The [1] PR will change the standard so that plugin documents
are in standard (no quirks) mode. This CL adds a test, as
much as this is testable. Some UAs display PDF documents
using "content that uses plugins" [2] while others use
"inline content that doesn't have a DOM [3]. This test will
pass in either case.

Chromium already made this change in code in Sept 2020
(see CLs on crbug.com/1131185) and found no compat
issues.

[1] https://github.com/whatwg/html/pull/6745
[2] https://html.spec.whatwg.org/#read-plugin
[3] https://html.spec.whatwg.org/#read-ua-inline

Bug: 1131185
Change-Id: Iccd1668199ac8b3b1c90a69b97a0adbaf8d02655