[selectmenu] Support the 'open' attribute
Add support for a readonly 'open' attribute on <selectmenu>, as lightly
specified by https://open-ui.org/components/select. This should be a
useful attribute to prioritize since it will make it more convenient to
write tests that modify the select listbox's open/closed state.

When adding test coverage I noticed an issue where default parts are
not removed properly when their <slot> is no longer showing them as the
fallback content. Fixing this may require some work in how we handle
part changes, so I plan to tackle that in a separate change.

Bug: 1121840
Change-Id: Ia92fba8857afd3f15e87e0392d1040377794eac8